requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
httpx==0.26.0
python-dotenv==1.0.0
google-genai==0.2.2